// src/teamcity-client/models/projects.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamCity REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2018.1 (current)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// May contain unused imports in some cases
// @ts-ignore
import type { Project } from './project';

/**
 * Represents a paginated list of Project entities.
 * @export
 * @interface Projects
 */
export interface Projects {
  /**
   * Returns the number of Project objects in this collection.
   * @type {number}
   * @memberof Projects
   */
  count?: number;
  /**
   * The relative link (without the server URL) to the current collection.
   * @type {string}
   * @memberof Projects
   */
  href?: string;
  /**
   * If the list of returned entities exceeds the request `count` value, TeamCity splits it into multiple batches. This property returns the endpoint that allows you to obtain the next batch.
   * @type {string}
   * @memberof Projects
   */
  nextHref?: string;
  /**
   * If the list of returned entities exceeds the request `count` value, TeamCity splits it into multiple batches. This property returns the endpoint that allows you to obtain the next batch.
   * @type {string}
   * @memberof Projects
   */
  prevHref?: string;
  /**
   * The list of Project entities owned by this collection.
   * @type {Array<Project>}
   * @memberof Projects
   */
  project?: Array<Project>;
}